added gallery and gallery filter

+++ app/controllers/PhotoController.php
@@ -12,14 +12,46 @@ class PhotoController extends Controller
 {
     public function listAction()
     {
-        print_r($this->cookies);
-        exit();
         $dataSet = DB::getInstance()->select("SELECT `path`, `name`, `id` FROM `photo`");
         $links = [];
         foreach ($dataSet as $data) {
             $links[] = [
-                "path" => $this->config["domain"] . $data["path"] ."min/". $data["name"],
+                "path" => $this->config["domain"] . $data["path"] . "min/" . $data["name"],
                 "id" => $data["id"]
+            ];
+        }
+        $this->response->setStatus("success");
+        $this->response->setMessage("OK");
+        $this->response->setData($links);
+        return $this->response->json();
+    }
+
+    public function photoAction()
+    {
+        $params = $this->request->getGet();
+        $data = [];
+        $where = "";
+        $lang = "en";
+        if (!empty($params["category"])){
+            $where = " LEFT JOIN photo_category pc on photo.id = pc.id_photo WHERE pc.id_category = :category";
+            $data["category"]=$params["category"];
+        }
+        $dataSet = DB::getInstance()->select(
+            "SELECT `path`, `name`, photo.id, title_{$lang} as title, description_{$lang} as description
+                    FROM `photo` 
+                    {$where}", $data
+        );
+        $links = [];
+        foreach ($dataSet as $data) {
+            $links[] = [
+                "path" => $this->config["domain"] ,
+                "name" => $data["name"],
+                "id" => $data["id"],
+                "min" => $data["path"] . "min/",
+                "middle" => $data["path"] . "middle/",
+                "full" => $data["path"],
+                "title" => $data["title"],
+                "description" => $data["description"]
             ];
         }
         $this->response->setStatus("success");
@@ -35,20 +67,9 @@ class PhotoController extends Controller
         $uploadFile = $uploadDir . $newFile;
 
         if (move_uploaded_file($_FILES['file']['tmp_name'], $uploadFile)) {
-            $minFile = new Imagick($uploadFile);
-            $imageprops = $minFile->getImageGeometry();
-            $width = $imageprops["width"];
-            $height = $imageprops["height"];
-            if ($height > $width) {
-                $newHeight = 240;
-                $newWidth = round(($width*$newHeight) / $height);
+            $this->resize($uploadFile, $uploadDir, $newFile, 320, 240, "min/");
+            $this->resize($uploadFile, $uploadDir, $newFile, 640, 480, "middle/");
 
-            } else {
-                $newWidth = 320;
-                $newHeight = round(($height / $width) * $newWidth);
-            }
-            $minFile->resizeImage($newWidth,$newHeight, Imagick::FILTER_LANCZOS, 1);
-            $minFile->writeImage($uploadDir . "min/" . $newFile);
             $id = DB::getInstance()->insert(
                 "INSERT INTO `photo` ( path, name, title_ua, title_en, description_ua, description_en) VALUES (:path, :name, :title_ua, :title_en, :description_ua, :description_en)",
                 [
@@ -100,6 +121,22 @@ class PhotoController extends Controller
         return $this->response->json();
     }
 
+    public function resize($uploadFile, $uploadDir, $newFile, $newWidth, $newHeight, $path)
+    {
+        $minFile = new Imagick($uploadFile);
+        $imageprops = $minFile->getImageGeometry();
+        $width = $imageprops["width"];
+        $height = $imageprops["height"];
+        if ($height > $width) {
+            $newWidth = round(($width * $newHeight) / $height);
+
+        } else {
+            $newHeight = round(($height / $width) * $newWidth);
+        }
+        $minFile->resizeImage($newWidth, $newHeight, Imagick::FILTER_LANCZOS, 1);
+        $minFile->writeImage($uploadDir . $path . $newFile);
+    }
+
     public function deleteAction()
     {
         $db = DB::getInstance();
@@ -115,7 +152,8 @@ class PhotoController extends Controller
             $this->response->setMessage("Photo not found");
         } else {
             if (unlink(BASE_PATH . "public" . $photoData[0]["path"] . $photoData[0]["name"])
-                && unlink(BASE_PATH . "public" . $photoData[0]["path"] . "min/" . $photoData[0]["name"])) {
+                && unlink(BASE_PATH . "public" . $photoData[0]["path"] . "min/" . $photoData[0]["name"])
+                && unlink(BASE_PATH . "public" . $photoData[0]["path"] . "middle/" . $photoData[0]["name"])) {
                 $db->delete('DELETE FROM `photo_category` WHERE `id_photo` = :id',
                     [
                         "id" => $this->request->getPost("id")
@@ -153,10 +191,10 @@ class PhotoController extends Controller
             [
                 "visible" => $flags[0]["visible"],
                 "slider" => $flags[0]["slider"],
-                "title_ua" =>$flags[0]["title_ua"],
-                "title_en" =>$flags[0]["title_en"],
-                "description_en" =>$flags[0]["description_en"],
-                "description_ua" =>$flags[0]["description_ua"],
+                "title_ua" => $flags[0]["title_ua"],
+                "title_en" => $flags[0]["title_en"],
+                "description_en" => $flags[0]["description_en"],
+                "description_ua" => $flags[0]["description_ua"],
                 "categories" => array_column($categories, "id_category")
             ]
         );
